refactor(seahaven-cli): move result error handling module
- Added a new `result` module to define a custom `Result` type and an `Error` struct for improved error handling in CLI commands.
- Updated existing command modules to use the new `Result` type from the `result` module instead of the previous import method.
