+++ seahaven-cli/src/main.rs
@@ -1,4 +1,5 @@
 mod cmd;
+mod result;
 
 #[tokio::main(flavor = "current_thread")]
 pub async fn main() {
