+++ seahaven-cli/src/result.rs
